[refactor] #19 NavigationView 재사용할 수 있게 리팩토링 (뷰 1, 2)
